Move merge metrics into one module
